+++ src/sim/model/ModelClass.py
@@ -1,7 +1,5 @@
-from threading import Thread
-import threading
-import ctypes
 from sim.model.v_main_test import v_main
+from sim.controller.KillableThreadClass import KillableThread
 
 class Model:
     def __init__(self):
@@ -18,33 +16,6 @@ class RobotModel:
     def __init__(self, ip):
         self.ip = ip
         self.robotItem = None
-        self.thread = thread_with_exception(v_main, [ip])
+        self.thread = KillableThread(v_main, (ip))
 
         self.thread.start()
-
-class thread_with_exception(Thread):
-    def __init__(self, target, args):
-        Thread.__init__(self)
-        self.__target = target
-        self.__args = args
-             
-    def run(self):
-        # Target
-        self.__target(self.__args[0])
-          
-    def get_id(self):
- 
-        # returns id of the respective thread
-        if hasattr(self, '_thread_id'):
-            return self._thread_id
-        for id, thread in threading._active.items():
-            if thread is self:
-                return id
-  
-    def raise_exception(self):
-        thread_id = self.get_id()
-        res = ctypes.pythonapi.PyThreadState_SetAsyncExc(thread_id,
-              ctypes.py_object(SystemExit))
-        if res > 1:
-            ctypes.pythonapi.PyThreadState_SetAsyncExc(thread_id, 0)
-            print('Exception raise failure')
